workunit/pjd:  Use modified pjd with open test
We have a modified version of pjd that includes an open test
(open.24).  The test ensures that open with mode=1000 and
O_CREAT|O_WRONLY succeeds on a non-existent file.

The modifications to the workunit use the pexec teuthology task
instead of the workunit task to avoid having to commit minor,
teuthology specific changes to the ceph tree directly
(namely the pjd.sh script).  Other than using the modified pjd
tarball, the script remains the same.

Signed-off-by: Sam Lang <sam.lang@inktank.com>