rgw: add default_idle_timeout to allow override
Globally overriding the rgw idle_timeout is not possible because it it
needs to be done on a per client.0, client.1, etc. basis. Add the
default_idle_timeout key to the rgw config : it defaults to the
previously hardcoded default (30) and can be changed via the override.

The existing tasks that were previously overriding the idle_timeout on a
per client basis are changed to use the default_idle_timeout instead for
consistency and to allow a global override.

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit 54a7298cdd8d40aaf79615a1f580b66f83e48419)

Conflicts:
	machine_types/vps.yaml
	suites/rgw/verify/tasks/rgw_s3tests_multiregion.yaml
	suites/smoke/verify/tasks/rgw_s3tests.yaml
	suites/upgrade/dumpling-x/stress-split/9-workload/rgw-s3tests.yaml
	tasks/rgw.py