erasure-code: lrc plugin workload
Add a workload that uses the lrc erasure code plugin. Instead of adding
it to suites/rados/thrash-erasure-code/workloads, a new suite is created
at suites/rados/thrash-erasure-code-big because it needs more OSDs than
other erasure code plugins. The alternative would be to increase the
number of OSDs for all erasure code plugins, but that would needlessly
increase the resources requirements.

 * cluster/12-osds.yaml creates a 12 OSDs, 3 MONs cluster

 * thrash-erasure-code-big/thrashers/*.yaml are the same as
   thrash-erasure-code/thrashers/*.yaml except they require that at
   least 8 OSDs are in at all times (instead of 4) because lrc PGs with
   k=4, m=2, l=3 are undersized if they do not have 8 OSDs. It is
   possible that crush fails to map 8 OSDs when only 8 OSDs are
   available, but that must not disturb the workload because min_size is
   4.

http://tracker.ceph.com/issues/11666 Fixes: #11666

Signed-off-by: Loic Dachary <ldachary@redhat.com>