Adds support to put daemons in cgroups
This closes http://tracker.ceph.com/issues/12424

As part of the work for http://tracker.ceph.com/issues/12379 we need a
way to provide (relative) hardware determinism, which we address by
using cgroups. This commit extends the ceph task so that cgroups can be
created and mon/osd/mds daemons can be placed in them.